refactor: 그룹 공통 컴포넌트 common/ 디렉토리 이동
