+++ src/app/password/find/page.tsx
@@ -1,6 +1,6 @@
-import PageHeader from "@/components/Header";
-import PageTitle from "@/components/PageTitle";
 import PasswordForm from "@/components/PasswordForm";
+import PageHeader from "@/components/common/Header";
+import PageTitle from "@/components/common/PageTitle";
 
 export default function PasswordPage() {
   return (
