+++ src/components/password/PasswordForm.tsx
@@ -4,38 +4,38 @@ import { useRouter } from "next/navigation";
 
 import { useState } from "react";
 
-import DateIcon from "@/assets/date.svg";
 import PhoneIcon from "@/assets/phone.svg";
 import ProfileIcon from "@/assets/profile.svg";
 
-import FullButton from "@/components/FullButton";
+import AlertModal from "@/components/common/AlertModal";
+import FullButton from "@/components/common/FullButton";
 
 import { useCountdown } from "@/hooks/useCountdown";
 import { usePhoneValidation } from "@/hooks/usePhoneValidation";
 
-import { formatBirthInput } from "@/utils/formatBirth";
 import { formatPhone } from "@/utils/formatPhone";
 import { formatTime } from "@/utils/formatTime";
 import { phoneChangeHandler } from "@/utils/phoneHandlers";
 
-export default function ProfileForm() {
+export default function PasswordForm() {
   const [name, setName] = useState("");
   const [phone, setPhone] = useState("");
-  const [birth, setBirth] = useState("");
   const [code, setCode] = useState("");
-  const [focused, setFocused] = useState<
-    "name" | "birth" | "phone" | "code" | null
-  >(null);
+  const [focused, setFocused] = useState<"name" | "phone" | "code" | null>(
+    null,
+  );
 
   const [isVerified, setIsVerified] = useState(false);
+  const [modalType, setModalType] = useState<"success" | "error" | null>(null);
 
   const router = useRouter();
 
-  const { timeLeft, isRunning, start, stop } = useCountdown();
   const { isValidPhone } = usePhoneValidation(phone);
-
+  const { timeLeft, isRunning, start, stop } = useCountdown();
   const isCodeRequested = isRunning;
+
   const isNameFilled = name.trim().length > 0;
+
   const canRequestCode = isValidPhone;
   const isConfirmActive = isNameFilled && isVerified;
 
@@ -54,23 +54,45 @@ export default function ProfileForm() {
 
   const handleVerify = () => {
     if (!code || !isCodeRequested || timeLeft === 0) return;
-    if (code === "1234") {
+    const isSuccess = code === "1234";
+
+    if (isSuccess) {
       setIsVerified(true);
+      setModalType("success");
       stop();
+    } else {
+      setIsVerified(false);
+      setModalType("error");
+    }
+  };
+
+  const handleModalConfirm = () => {
+    if (modalType === "success") {
+      setModalType(null);
+    } else {
+      setModalType(null);
+      stop();
+      setCode("");
+      setIsVerified(false);
     }
   };
 
   const handlePhoneChange = phoneChangeHandler(setPhone);
-
-  const handleBirthChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setBirth(formatBirthInput(e.target.value));
+  const handlePhoneInput: React.ChangeEventHandler<HTMLInputElement> = e => {
+    handlePhoneChange(e);
+    // 번호를 바꾸면 인증 상태/타이머 리셋
+    if (isCodeRequested || timeLeft > 0 || isVerified) {
+      stop();
+      setIsVerified(false);
+      setCode("");
+    }
   };
 
   return (
-    <div className="mt-[18px] flex w-full flex-col gap-2.5 px-4">
+    <div className="mt-[75px] flex w-full flex-col gap-[17px] px-4">
       {/* 이름 */}
       <div
-        className={`bg-neutral-11 flex h-[55px] w-full items-center rounded-2xl border px-2.5 py-2 ${
+        className={`bg-neutral-11 flex h-[55px] w-full items-center rounded-2xl border px-[10px] py-[8px] ${
           focused === "name" || name.length > 0
             ? "border-mint-01"
             : "border-neutral-08"
@@ -80,7 +102,7 @@ export default function ProfileForm() {
           <ProfileIcon />
         </div>
         <input
-          type="text"
+          type="name"
           value={name}
           onChange={e => setName(e.target.value)}
           onFocus={() => setFocused("name")}
@@ -90,30 +112,9 @@ export default function ProfileForm() {
         />
       </div>
 
-      {/* 생년월일 */}
-      <div
-        className={`bg-neutral-11 flex h-[55px] w-full items-center rounded-2xl border px-2.5 py-2 ${
-          focused === "birth" || birth.length > 0
-            ? "border-mint-01"
-            : "border-neutral-08"
-        }`}
-      >
-        <div className="pr-2.5">
-          <DateIcon />
-        </div>
-        <input
-          type="text"
-          value={birth}
-          onChange={handleBirthChange}
-          onFocus={() => setFocused("birth")}
-          onBlur={() => setFocused(prev => (prev === "birth" ? null : prev))}
-          placeholder="생년월일 8자리"
-          className="placeholder:text-neutral-07 text-h2 w-full bg-transparent text-black outline-none"
-        />
-      </div>
-
-      {/* 전화번호 + 인증번호 */}
+      {/* 전화번호 + 인증번호 받기 */}
       <div className="flex w-full gap-[13px]">
+        {/* 전화번호 입력 */}
         <div
           className={`bg-neutral-11 flex h-[55px] flex-1 items-center gap-[12px] rounded-2xl border px-[10px] py-[8px] ${
             focused === "phone" || phone.length > 0
@@ -125,7 +126,7 @@ export default function ProfileForm() {
           <input
             type="tel"
             value={formatPhone(phone)}
-            onChange={handlePhoneChange}
+            onChange={handlePhoneInput}
             onFocus={() => setFocused("phone")}
             onBlur={() => setFocused(prev => (prev === "phone" ? null : prev))}
             placeholder="전화번호"
@@ -133,13 +134,14 @@ export default function ProfileForm() {
           />
         </div>
 
+        {/* 인증번호 받기 버튼 */}
         <button
           type="button"
           onClick={handleRequestCode}
-          disabled={!canRequestCode || isCodeRequested}
-          className={`text-h2 flex h-[55px] items-center justify-center rounded-2xl border px-[10px] py-[8px] whitespace-nowrap ${
-            !canRequestCode
-              ? "border-neutral-08 text-neutral-06 bg-white"
+          disabled={!isValidPhone || isCodeRequested}
+          className={`text-h2 flex h-[55px] cursor-pointer items-center justify-center rounded-2xl border px-[10px] py-[8px] whitespace-nowrap ${
+            !isValidPhone
+              ? "border-neutral-08 text-neutral-06 border bg-white"
               : isCodeRequested
                 ? "border-neutral-07 bg-neutral-07 text-white"
                 : "bg-mint-01 text-white"
@@ -151,6 +153,7 @@ export default function ProfileForm() {
 
       {/* 인증번호 + 인증하기 */}
       <div className="flex w-full gap-[13px]">
+        {/* 인증번호 입력 */}
         <div className="flex flex-1 items-center gap-[10px]">
           <div
             className={`bg-neutral-11 flex h-[55px] flex-1 items-center gap-[13px] rounded-2xl border px-[10px] py-[8px] ${
@@ -180,10 +183,10 @@ export default function ProfileForm() {
           type="button"
           onClick={handleVerify}
           disabled={!code || !isCodeRequested || timeLeft === 0}
-          className={`text-h2 flex h-[55px] w-[125px] items-center justify-center rounded-2xl border px-[10px] py-[8px] ${
+          className={`text-h2 flex h-[55px] w-[125px] cursor-pointer items-center justify-center rounded-2xl border px-[10px] py-[8px] ${
             !code || !isCodeRequested || timeLeft === 0
               ? "border-neutral-08 text-neutral-06 bg-white"
-              : "border-mint-01 bg-mint-01 cursor-pointer text-white"
+              : "border-mint-01 bg-mint-01 text-white"
           }`}
         >
           인증하기
@@ -195,21 +198,34 @@ export default function ProfileForm() {
           <button
             type="button"
             onClick={handleResendClick}
-            className="text-orange-00 text-[10px] underline"
+            className="text-orange-00 cursor-pointer text-[10px] leading-[150%] underline"
           >
             인증번호가 오지 않나요?
           </button>
         </div>
       )}
 
-      <div className="mt-[188px] flex w-full justify-center">
+      <div className="mt-[270px] flex w-full justify-center">
         <FullButton
           isActive={isConfirmActive}
-          onClick={() => router.push("/signup/password")}
+          onClick={() => router.push("/password/find/result")}
         >
           확인
         </FullButton>
       </div>
+
+      {modalType && (
+        <AlertModal
+          isOpen={!!modalType}
+          title={modalType === "success" ? "인증완료" : "인증오류"}
+          message={
+            modalType === "success"
+              ? "인증이 완료되었습니다."
+              : "인증번호를 확인할 수 없습니다."
+          }
+          onClose={handleModalConfirm}
+        />
+      )}
     </div>
   );
 }
